requirements.txt
# Bluetooth Low Energy communication with MELK-OA devices
bleak

# Web server for Nanoleaf API emulation
flask
flask-cors

# mDNS discovery (auto-detection in SignalRGB)
zeroconf

# Auto power-on functionality for MELK-OA devices
btledstrip

# System tray GUI
pystray
pillow

# GUI window (tkinter is usually included with Python, but listing for completeness)
# tkinter - usually included with Python

# Modern GUI styling (Material Design 3)
ttkbootstrap

# Build tools (optional, for creating executable)
pyinstaller
pyarmor